refactor: Improve x/bank deterministic tests
