+++ testutil/testdata/tx.go
@@ -11,6 +11,7 @@ import (
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/cosmos/cosmos-sdk/types/query"
 )
 
 // AddressGenerator creates and returns a random address generator using rapid.
@@ -18,6 +19,19 @@ func AddressGenerator(t *rapid.T) *rapid.Generator[sdk.AccAddress] {
 	return rapid.Custom(func(t *rapid.T) sdk.AccAddress {
 		pkBz := rapid.SliceOfN(rapid.Byte(), 20, 20).Draw(t, "hex")
 		return sdk.AccAddress(pkBz)
+	})
+}
+
+// AddressGenerator creates and returns a pagination PageRequest generator
+// using rapid.
+func PaginationGenerator(t *rapid.T, maxLimit uint64) *rapid.Generator[*query.PageRequest] {
+	return rapid.Custom(func(t *rapid.T) *query.PageRequest {
+		return &query.PageRequest{
+			Offset:     rapid.Uint64Range(0, maxLimit).Draw(t, "offset"),
+			Limit:      rapid.Uint64Range(0, maxLimit).Draw(t, "limit"),
+			CountTotal: rapid.Bool().Draw(t, "count-total"),
+			Reverse:    rapid.Bool().Draw(t, "reverse"),
+		}
 	})
 }
